Discount & Checkout Working functionally

+++ Models/Order.cs
@@ -6,7 +6,7 @@ public class Order {
     public DateTime RequiredDate { get; set; }
     public DateTime ShippedDate { get; set; }
     public int ShipVia { get; set; }
-    public double Freight { get; set; }
+    public decimal Freight { get; set; }
     public string ShipName { get; set; }
     public string ShipAddress { get; set; }
     public string ShipCity { get; set; }
